Update Form-Checkboxes.md
Small changes to English punctuation and spelling.

Questions:

 On line 60. "The accessible name comes the child text content of the div[role="checkbox"] element." is incorrect English. The verb "comes" is used wrongly. What exactly is meant, here? 

A similar issue to the above exists, on line 69.

On line 72, "aria-checked="false" | div | Identifies the checkbox button as not unchecked." - is this incorrect? Should it not read "aria-checked="false" | div | Identifies the checkbox button as not checked."?

Line 82 seems, to me, to contradict line 87.