Update Navigation-Pagination.md
Small punctuation and spelling changes.

On line 5, I changed "performant matter" to "performance-efficient manner". The word 'performant' is a disputed computer term. In plain English, a 'performant' is the one who performs. The word 'matter' is almost certainly meant to be 'manner'.

On line 8, I changed 'pattern' to 'subject', as I think that that is the more correct word to use. I may be wrong.

On line 23, the colon in '**Space:**' is probably unnecessary. Should be '**space**' (??)

On line 69, I have changed the grammatical structure, considerably.