Update understanding-disability.md
Minor changes to punctuation and grammar.

Line 106 uses the word "strightly". This is not a correct English word, but I cannot work out what is intended. Should it be "strictly", "rightly" or something else. Perhaps, the word is not needed, at all or that it should be something very different, such as "often".